# api.yaml
openapi: 3.0.0
servers:
  - url: http://localhost:8080
info:
  title: CIRCABC Share
  description: This is a API definition for the CIRCABC Share service.
  license:
    name: 'EUPL V1.2'
    url: 'https://joinup.ec.europa.eu/sites/default/files/inline-files/EUPL%20v1_2%20EN(1).txt'
  version: '0.1'
...
paths:
...
  '/file/{fileID}':
    parameters:
    - name: fileID
      in: path
      description: The id of the file
      required: true
      schema:
        type: string
    get:
      description: Used by INTERNAL and EXTERNAL users to download a shared file
      operationId: getFile
      parameters:
        - name: password
          in: query
          description: 'Password of the file'
          required: false
          schema:
            type: string
            format: base64
      tags:
      - File
      responses:
        200:
          description: SUCCESS Returns the file and sends a mail to original uploader to inform him of the download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
    delete:
      description: Used by INTERNAL users and ADMIN in order to delete a file
      operationId: deleteFile
      parameters:
        - name: reason
          in: query
          description: 'Reason for deletion of the file'
          required: false
          schema:
            type: string
      tags:
      - File
      security:
        - OpenID:
          - email
      responses:
        200:
          description: SUCCESS Deletes the file content and its meta data, also sends an email to original uploader with the reason of deletion if the reason is given
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'   
    put:
      description: Used by ADMIN to update file's metadata (expiration date)
      operationId: updateFile
      tags:
      - File
      security:
        - OpenID:
          - email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileBasics'
      responses:
        200:
          description: SUCCESS Updates the file content and its meta data
        401:
          description: UNAUTHORIZED the Error message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        403:
          description: FORBIDDEN the Error message will be NotAuthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        404:
          description: NOT FOUND the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        500:
          description: INTERNAL SERVER ERROR the Error Message will be empty
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status' 
...
  '/abuse':
    post:
      tags:
      - Abuse
      summary: Create a new abuse report
      operationId: "createAbuseReport"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbuseReport"
      
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbuseReport"
    get:
      tags:
        - Abuse
      summary: Get a list of abuse reports with details grouped by fileId
      operationId: getAbuseReportList
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/AbuseReportDetails'
  '/abuse/{id}':
    get:
      tags:
      - Abuse
      summary: Get a specific abuse report by ID.
      operationId: "getAbuseReport"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbuseReport'
    put:
      tags:
      - Abuse
      summary: Update a specific abuse report by ID
      operationId: "updateAbuseReport"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbuseReportDetails'
      responses:
        '200':
          description: "Successful Operation"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbuseReportDetails"
    delete:
      tags:
      - Abuse
      summary: Delete a specific abuse report by ID
      operationId: "deleteAbuseReport"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: No Content
...
components:
  securitySchemes:
    OpenID:   # <--- Arbitrary name for the security scheme. Used to refer to it from elsewhere.
      type: openIdConnect
      openIdConnectUrl: https://localhost:8080/realms/
...
  schemas:
    Status:
      type: object
      properties:
        code:
          type: integer
          description: Status code
        message:
          type: string
          description: Additional message
      required:
        - code
...
    FileBasics:
      type: object
      properties:
        expirationDate:
          description: Expiration date of file
          type: string
          format: date
        hasPassword:
          description: File is password-protected
          type: boolean
        name:
          description: Filename
          type: string
        size:
          description: Size of file (Bytes)
          type: number
          minimum: 0
      required:
      - expirationDate
      - hasPassword
      - name
      - size
...
    AbuseReport:
      type: object
      properties:
        ID:
          type: string
        reporter:
          type: string
        fileId:
          type: string
        reason:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
          enum:
            - WAITING
            - DENIED
            - APPROVED
    AbuseReportDetails:
      allOf:
        - $ref: '#/components/schemas/AbuseReport'
        - type: object
          properties:
            filename:
              type: string
            filesize:
              type: number
            shortUrl:
              type: string
            uploader_email:
              type: string
            uploader_name:
              type: string
            uploader_status:
              type: string
...